+++ src/main.ts
@@ -1,2 +1,3 @@
 export { default as Slider } from './utilities/slider';
 export { default as FormValidation } from './utilities/form-validation';
+export { default as StickyHeaders } from './utilities/sticky-headers';